Commit 18

+++ app/src/main/java/com/example/android/workout/Exercise.java
@@ -30,7 +30,7 @@ public class Exercise implements Parcelable{
         mMuscleGroup = in.readString();
         mType = in.readString();
         mEquipment = in.readString();
-        mWeight = in.readInt();
+        mWeight = in.readFloat();
         mReps = in.readInt();
     }
